+++ backend_routes/vcontrol.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const express = require('express');
+const router = express.Router();
+const Models = require('../db/models');
+const Actions = require('../utilities/actions');
+
+router.get('/', (req, res, next) => {
+    
+})
